move run_query
