+++ crates/iota-sdk-ffi/src/graphql/client.rs
@@ -3,7 +3,7 @@
 
 use tokio::sync::RwLock;
 
-use crate::error::Result;
+use crate::error::{Result, SdkFfiError};
 
 /// The GraphQL client for interacting with the IOTA blockchain.
 #[derive(uniffi::Object)]
@@ -17,6 +17,14 @@ impl GraphQLClient {
     pub fn into_inner(self) -> RwLock<iota_sdk::graphql_client::Client> {
         self.0
     }
+}
+
+#[derive(Debug, uniffi::Record, serde::Serialize)]
+pub struct Query {
+    pub query: String,
+    #[uniffi(default = None)]
+    #[serde(default)]
+    pub variables: Option<serde_json::Value>,
 }
 
 #[uniffi::export(async_runtime = "tokio")]
@@ -66,5 +74,21 @@ impl GraphQLClient {
     /// valid URL with a host and optionally a port number.
     pub async fn set_rpc_server(&self, server: String) -> Result<()> {
         Ok(self.0.write().await.set_rpc_server(&server)?)
+    }
+
+    /// Run a query.
+    pub async fn run_query(&self, query: Query) -> Result<serde_json::Value> {
+        self.0
+            .read()
+            .await
+            .run_query_from_json(
+                serde_json::to_value(query)?
+                    .as_object()
+                    .ok_or_else(|| SdkFfiError::custom("invalid json; must be a map"))?
+                    .clone(),
+            )
+            .await?
+            .data
+            .ok_or_else(|| SdkFfiError::custom("query yielded no data"))
     }
 }
